+++ CHANGELOG.md
@@ -16,6 +16,7 @@ HumHub Changelog
 - Fix #6674: Fix visibility of draft and scheduled content on dashboard
 - Fix #6679: On the space invitation by link page, the language chooser doesn't save the selected language
 - Fix #6677: Do not check mobile push module in requirements check on installation
+- Fix #6694: Remove the "Offline" tooltip when my own "Online" status is not displayed
 - Fix #6700: Reload theme variable `@baseTheme` after flushing cache
 
 1.15.0 (November 6, 2023)